Put various metrics and plots in individual files
